standardize random number seed parameters in c++ source code to use unsigned long long type uniformly

+++ src/GridExp.cpp
@@ -2175,7 +2175,7 @@ Rcpp::DataFrame RcppGPCRobust4Grid(
     int b = 4,
     int boot = 99,
     bool random = true,
-    unsigned int seed = 42,
+    unsigned long long seed = 42,
     int zero_tolerance = 0,
     int dist_metric = 2,
     bool relative = true,
